+++ db/migrations/20190228121351_create_hash_released.sql
@@ -3,7 +3,7 @@ CREATE TABLE ens.hash_released (
   id                SERIAL PRIMARY KEY,
   header_id         INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
   hash              CHARACTER VARYING(66) NOT NULL,
-  value             BIGINT NOT NULL,
+  value             NUMERIC NOT NULL,
   tx_idx            INTEGER NOT NUll,
   log_idx           INTEGER NOT NUll,
   raw_log           JSONB,
