+++ db/migrations/20190228121529_create_abi_changed.sql
@@ -3,7 +3,7 @@ CREATE TABLE ens.abi_changed (
   id                SERIAL PRIMARY KEY,
   header_id         INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
   node              CHARACTER VARYING(66) NOT NULL,
-  content_type      BIGINT NOT NULL,
+  content_type      NUMERIC NOT NULL,
   tx_idx            INTEGER NOT NUll,
   log_idx           INTEGER NOT NUll,
   raw_log           JSONB,
